Add shared types header
